Refactor anomaly detection to be more generic
