+++ anomaly-detection/esd.py
@@ -2,21 +2,29 @@
 Provides a Web Service that does Anomaly Detection.
 
 Parameters are:
-- name of mongod-db collection, where the Tweets to be modelled are
-- unix-timestamp with start of time range to be included
-- unix-timestamp with stop of time range to be included
+- List of values of a timeseries
+- Frequence of the timeseries
+- Desired p value
 
-Example-Call:
-http://0.0.0.0:5001/esd?collection=bitcoin&start=1515082825836&end=1515082840114
+Example BODY json/application:
+{
+"ary": [112, 118, 132, 129, 121, 135, 148, 148, 136, 119, 104, 118, 115, 126,
+ 141, 135, 125, 149, 170, 170, 158, 133, 114, 140, 145, 150, 178, 163, 172,
+ 178, 199, 199, 184, 162, 146, 166, 171, 180, 193, 181, 183, 218, 230, 242,
+ 209, 191, 172, 194, 196, 196, 236, 235, 229, 243, 264, 272, 237, 211, 180,
+ 201, 204, 188, 235, 227, 234, 264, 302, 293, 259, 229, 203, 229, 242, 233,
+ 267, 269, 270, 315, 364, 347, 312, 274, 237, 278, 284, 277, 317, 313, 318,
+ 374, 413, 405, 355, 306, 271, 306, 315, 301, 356, 348, 355, 422, 465, 467,
+ 404, 347, 305, 336, 340, 318, 362, 348, 363, 435, 491, 505, 404, 359, 310,
+ 337, 360, 342, 406, 396, 420, 472, 548, 559, 463, 407, 362, 405, 417, 391,
+ 419, 461, 472, 535, 622, 606, 508, 461, 390, 432],
+"freq": 12,
+"p": 0.95
+}
 """
 
-import pandas as pd
 import json
-import yaml
-from pymongo import MongoClient
-from flask import Flask
-from webargs import fields
-from webargs.flaskparser import use_args
+from flask import Flask, abort, request
 import statsmodels.api as sm
 import numpy as np
 from PyAstronomy import pyasl
@@ -27,128 +35,76 @@ logging.basicConfig(format='%(levelname)s - %(asctime)s: %(message)s',
 logger = logging.getLogger(__name__)
 
 
-def agg_tweets(db, collection, start_time, end_time, hours):
-    """Load text from tweets in specified colleciton & time range."""
-    agg_range = 1000 * 60 * 60 * hours   # by hour
-    cursor = db[collection].aggregate([
-        {
-            '$project': {
-                'timestamp_ms': '$timestamp_ms',
-                'div_val': {'$divide': ['$timestamp_ms', agg_range]},
-            }
-        },
-        {
-            '$project': {
-                'timestamp_ms': '$timestamp_ms',
-                'div_val': '$div_val',
-                'mod_val': {'$mod': ['$div_val', 1]}
-            }
-        },
-        {
-            '$project': {
-                'timestamp_ms': '$timestamp_ms',
-                'div_val': '$div_val',
-                'mod_val': '$mod_val',
-                'sub_val': {'$subtract': ['$div_val', '$mod_val']},
-            }
-        },
-        {
-            '$group': {
-                '_id': '$sub_val',
-                'count': {'$sum': 1}
-            }
-        }])
-    df = pd.DataFrame(list(cursor))
-    df['collection'] = collection
-
-    # Restore the actual time stamps, which got "compressed"
-    # during mongodb aggregation
-    df['timestamp_ms'] = df['_id'].astype(int).multiply(agg_range)
-
-    df = df[(df['timestamp_ms'] >= start_time)
-            & (df['timestamp_ms'] <= end_time)]
-
-    # Convert to datetime
-    df['date'] = pd.to_datetime(df['timestamp_ms'], unit='ms')
-
-    # Set as index
-    df = df.set_index(df['date'])
-    df = df.sort_index()
-
-    # Remove the mongo-row-id, as it's not needed
-    if '_id' in df.columns:
-        del df['_id']
-    if 'date' in df.columns:
-        del df['date']
-    return df
-
-
-def detect_anomalies(df, hours):
+def detect_anomalies(ary, freq, p):
     """Use Seasonal Decompose and ESD on residual to detect anomalies."""
-    # The frequency of the signal changes. We expect 24h cycles.
-    frequency = round(24 / hours)
-
     # Seasonal decompose
-    model = sm.tsa.seasonal_decompose(df['count'], freq=frequency)
+    model = sm.tsa.seasonal_decompose(ary, freq=freq)
 
     # We only use residue values here
-    resid = model.resid.values
+    resid = model.resid
 
-    # Set NaNs to 0, as this ESD doesn't handle them
-    resid[np.isnan(resid)] = 0
+    # Count leading Nan, which are going to be dropped for ESD
+    # but have to be taken into account for the results
+    dropped = 0
+    for val in resid:
+        if np.isnan(val):
+            dropped += 1
+        else:
+            break
+
+    # Remove NaNs, as ESD doesn't handle them well
+    resid_cleaned = [x for x in resid if not np.isnan(x)]
 
     # Use ESD to detect anomalies
-    anomalies = pyasl.generalizedESD(resid, 10, 0.05)
+    anomalies = pyasl.generalizedESD(resid_cleaned, 20, p)
 
-    # Get the timestamps from the corresponding anomalies
-    result_df = df['timestamp_ms'].ix[anomalies[1]].tolist()
-    return result_df
+    # Get the indexed of the anomalies (index of resid_cleaned + dropped)
+    idx_anoms = [x + dropped for x in anomalies[1]]
 
+    # Convert np.int64 to standard int to make it serializable in json
+    idx_anoms = [x.item() for x in idx_anoms]
 
-def open_mongo():
-    """Open Connection to MongoDB and return db-object."""
-    # Load config yaml
-    with open('../config.yaml', 'r') as stream:
-        config = yaml.load(stream)
-
-    # Open Connection to MongoDB
-    conn = MongoClient(config['mongodb']['host'],
-                       config['mongodb']['port'])
-    # Use local mongo-container IP for testing
-    conn = MongoClient('172.17.0.2', config['mongodb']['port'])
-    db = conn[config['mongodb']['db']]
-    return db
+    return idx_anoms, len(resid_cleaned)
 
 
 def init_flask():
     """Initialize Flask Webservice."""
-    # Open MongoDB needed for answering request
-    db = open_mongo()
-
     # Define Flask Webservice
     app = Flask(__name__)
 
-    anom_args = {
-        'collection': fields.Str(missing='bitcoin'),
-        'start': fields.Int(missing=0),
-        'end': fields.Int(missing=0),
-        'hours': fields.Int(missing=1),
-    }
-
-    @app.route('/esd', methods=['GET'])
-    @use_args(anom_args)
-    def detect_anoms(args):
+    @app.route('/esd', methods=['POST'])
+    def detect_anoms():
         """Handle incoming request, send back anomalies."""
         logger.info('Received request.')
-        result = {}
-        df = agg_tweets(db, args['collection'], args[
-                        'start'], args['end'], args['hours'])
+        print(request)
+        if not request.json:
+            logger.warn('No valid JSON in Request.')
+            abort(400)
 
-        result['agg_hours'] = args['hours']
-        result['values_count'] = len(df)
-        result['anomalies'] = detect_anomalies(df, args['hours'])
-        logger.info('Sending result:', result)
-        return json.dumps(result)
+        data = request.get_json()
+
+        # Sanitize Request
+        if 'ary' not in data:
+            abort(400)
+            logger.warn('No data sent.')
+        if ('freq' not in data):
+            data['freq'] = 24
+            logger.info('Using default freqency of 24.')
+        if ('p' not in data):
+            data['p'] = 0.05
+            logger.info('Using default freqency of 0.05.')
+
+        # Build Response
+        res = {}
+        res['p'] = data['p']
+        res['freq'] = data['freq']
+        res['values_received'] = len(data['ary'])
+        res['idx_anoms'], res['values_used'] = detect_anomalies(data['ary'],
+                                                                data['freq'],
+                                                                data['p'])
+        logger.info('Sending result:', res)
+        print(res)
+        return json.dumps(res)
     return app
